Updated RunSqlWithRollBack  sub

+++ IspDataDb/dbo/Tables/PurchaseDetail.sql
@@ -3,8 +3,8 @@
     [Sequence]       SMALLINT        NULL,
     [PurchaseIdNo]   INT             NULL,
     [ProductIdNo]    INT             NULL,
-    [Quantity]       SMALLINT        NULL,
-    [BonusQuantity]  SMALLINT        NULL,
+    [Quantity]       DECIMAL (12, 4) NULL,
+    [BonusQuantity]  DECIMAL (12, 4) NULL,
     [UnitIdNo]       TINYINT         NULL,
     [BatchNo]        VARCHAR (20)    NULL,
     [Price]          DECIMAL (9, 2)  NULL,
@@ -17,6 +17,8 @@
     [ExtraDiscount]  DECIMAL (5, 2)  NULL,
     CONSTRAINT [PK_PurchaseDetail] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
